Main moves of all pieces implemented. Now The king and queen can move correctly, and the rook has been fixed
Now there are some special rules to be implemented such as, castling, and check. If the king is moving it can not move to a square controlled by some other opponent's pieces. This is something to fix. Also if the to king are next to one another, the can't eat each other, or be next to each other. There must always be at least a square of distance between them.
#ToDo: other pieces can not eat the opponent's king: so the king square must be removed and the check function (to implement) called